+++ endpoints/publicProxy/http.go
@@ -2,11 +2,14 @@ package publicProxy
 
 import (
 	"context"
+	"crypto/md5"
 	"fmt"
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/openziti/sdk-golang/ziti"
 	"github.com/openziti/zrok/endpoints"
 	"github.com/openziti/zrok/endpoints/publicProxy/healthUi"
 	"github.com/openziti/zrok/endpoints/publicProxy/notFoundUi"
+	"github.com/openziti/zrok/endpoints/publicProxy/unauthorizedUi"
 	"github.com/openziti/zrok/environment"
 	"github.com/openziti/zrok/sdk"
 	"github.com/openziti/zrok/util"
@@ -17,20 +20,34 @@ import (
 	"net/http/httputil"
 	"net/url"
 	"strings"
+	"time"
 )
 
-type httpFrontend struct {
+type HttpFrontend struct {
 	cfg     *Config
 	zCtx    ziti.Context
 	handler http.Handler
 }
 
-func NewHTTP(cfg *Config) (*httpFrontend, error) {
-	env, err := environment.LoadRoot()
+func NewHTTP(cfg *Config) (*HttpFrontend, error) {
+	var key []byte
+	if cfg.Oauth != nil {
+		hash := md5.New()
+		n, err := hash.Write([]byte(cfg.Oauth.HashKey))
+		if err != nil {
+			return nil, err
+		}
+		if n != len(cfg.Oauth.HashKey) {
+			return nil, errors.New("short hash")
+		}
+		key = hash.Sum(nil)
+	}
+
+	root, err := environment.LoadRoot()
 	if err != nil {
 		return nil, errors.Wrap(err, "error loading environment root")
 	}
-	zCfgPath, err := env.ZitiIdentityNamed(cfg.Identity)
+	zCfgPath, err := root.ZitiIdentityNamed(cfg.Identity)
 	if err != nil {
 		return nil, errors.Wrapf(err, "error getting ziti identity '%v' from environment", cfg.Identity)
 	}
@@ -52,26 +69,28 @@ func NewHTTP(cfg *Config) (*httpFrontend, error) {
 		return nil, err
 	}
 	proxy.Transport = zTransport
-
-	handler := authHandler(util.NewProxyHandler(proxy), "zrok", cfg, zCtx)
-	return &httpFrontend{
+	if err := configureOauthHandlers(context.Background(), cfg, false); err != nil {
+		return nil, err
+	}
+	handler := authHandler(util.NewProxyHandler(proxy), cfg, key, zCtx)
+	return &HttpFrontend{
 		cfg:     cfg,
 		zCtx:    zCtx,
 		handler: handler,
 	}, nil
 }
 
-func (self *httpFrontend) Run() error {
-	return http.ListenAndServe(self.cfg.Address, self.handler)
+func (f *HttpFrontend) Run() error {
+	return http.ListenAndServe(f.cfg.Address, f.handler)
 }
 
 type zitiDialContext struct {
 	ctx ziti.Context
 }
 
-func (self *zitiDialContext) Dial(_ context.Context, _ string, addr string) (net.Conn, error) {
+func (c *zitiDialContext) Dial(_ context.Context, _ string, addr string) (net.Conn, error) {
 	shrToken := strings.Split(addr, ":")[0] // ignore :port (we get passed 'host:port')
-	conn, err := self.ctx.Dial(shrToken)
+	conn, err := c.ctx.Dial(shrToken)
 	if err != nil {
 		return conn, err
 	}
@@ -128,9 +147,9 @@ func hostTargetReverseProxy(cfg *Config, ctx ziti.Context) *httputil.ReverseProx
 	return &httputil.ReverseProxy{Director: director}
 }
 
-func authHandler(handler http.Handler, realm string, cfg *Config, ctx ziti.Context) http.HandlerFunc {
+func authHandler(handler http.Handler, pcfg *Config, key []byte, ctx ziti.Context) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		shrToken := resolveService(cfg.HostMatch, r.Host)
+		shrToken := resolveService(pcfg.HostMatch, r.Host)
 		if shrToken != "" {
 			if svc, found := endpoints.GetRefreshedService(shrToken, ctx); found {
 				if cfg, found := svc.Config[sdk.ZrokProxyConfig]; found {
@@ -145,7 +164,7 @@ func authHandler(handler http.Handler, realm string, cfg *Config, ctx ziti.Conte
 							logrus.Debugf("auth scheme basic '%v", shrToken)
 							inUser, inPass, ok := r.BasicAuth()
 							if !ok {
-								writeUnauthorizedResponse(w, realm)
+								basicAuthRequired(w, shrToken)
 								return
 							}
 							authed := false
@@ -179,15 +198,96 @@ func authHandler(handler http.Handler, realm string, cfg *Config, ctx ziti.Conte
 							}
 
 							if !authed {
-								writeUnauthorizedResponse(w, realm)
+								basicAuthRequired(w, shrToken)
 								return
 							}
 
 							handler.ServeHTTP(w, r)
 
+						case string(sdk.Oauth):
+							logrus.Debugf("auth scheme oauth '%v'", shrToken)
+
+							if oauthCfg, found := cfg["oauth"]; found {
+								if provider, found := oauthCfg.(map[string]interface{})["provider"]; found {
+									var authCheckInterval time.Duration
+									if checkInterval, found := oauthCfg.(map[string]interface{})["authorization_check_interval"]; !found {
+										logrus.Errorf("Missing authorization check interval in share config. Defaulting to 3 hours")
+										authCheckInterval = 3 * time.Hour
+									} else {
+										i, err := time.ParseDuration(checkInterval.(string))
+										if err != nil {
+											logrus.Errorf("unable to parse authorization check interval in share config (%v). Defaulting to 3 hours", checkInterval)
+											authCheckInterval = 3 * time.Hour
+										} else {
+											authCheckInterval = i
+										}
+									}
+
+									target := fmt.Sprintf("%s%s", r.Host, r.URL.Path)
+
+									cookie, err := r.Cookie("zrok-access")
+									if err != nil {
+										logrus.Errorf("unable to get 'zrok-access' cookie: %v", err)
+										oauthLoginRequired(w, r, shrToken, pcfg, provider.(string), target, authCheckInterval)
+										return
+									}
+									tkn, err := jwt.ParseWithClaims(cookie.Value, &ZrokClaims{}, func(t *jwt.Token) (interface{}, error) {
+										if pcfg.Oauth == nil {
+											return nil, fmt.Errorf("missing oauth configuration for access point; unable to parse jwt")
+										}
+										return key, nil
+									})
+									if err != nil {
+										logrus.Errorf("unable to parse jwt: %v", err)
+										oauthLoginRequired(w, r, shrToken, pcfg, provider.(string), target, authCheckInterval)
+										return
+									}
+									claims := tkn.Claims.(*ZrokClaims)
+									if claims.Provider != provider {
+										logrus.Error("provider mismatch; restarting auth flow")
+										oauthLoginRequired(w, r, shrToken, pcfg, provider.(string), target, authCheckInterval)
+										return
+									}
+									if claims.AuthorizationCheckInterval != authCheckInterval {
+										logrus.Error("authorization check interval mismatch; restarting auth flow")
+										oauthLoginRequired(w, r, shrToken, pcfg, provider.(string), target, authCheckInterval)
+										return
+									}
+									if validDomains, found := oauthCfg.(map[string]interface{})["email_domains"]; found {
+										if castedDomains, ok := validDomains.([]interface{}); !ok {
+											logrus.Error("invalid email domain format")
+											return
+										} else {
+											if len(castedDomains) > 0 {
+												found := false
+												for _, domain := range castedDomains {
+													if strings.HasSuffix(claims.Email, domain.(string)) {
+														found = true
+														break
+													}
+												}
+												if !found {
+													logrus.Warnf("invalid email domain")
+													unauthorizedUi.WriteUnauthorized(w)
+													return
+												}
+											}
+										}
+									}
+									handler.ServeHTTP(w, r)
+									return
+
+								} else {
+									logrus.Warnf("%v -> no provider for '%v'", r.RemoteAddr, provider)
+									notFoundUi.WriteNotFound(w)
+								}
+							} else {
+								logrus.Warnf("%v -> no oauth cfg for '%v'", r.RemoteAddr, shrToken)
+								notFoundUi.WriteNotFound(w)
+							}
 						default:
 							logrus.Infof("invalid auth scheme '%v'", scheme)
-							writeUnauthorizedResponse(w, realm)
+							basicAuthRequired(w, shrToken)
 							return
 						}
 					} else {
@@ -209,14 +309,53 @@ func authHandler(handler http.Handler, realm string, cfg *Config, ctx ziti.Conte
 	}
 }
 
-func writeUnauthorizedResponse(w http.ResponseWriter, realm string) {
+type ZrokClaims struct {
+	Email                      string        `json:"email"`
+	AccessToken                string        `json:"accessToken"`
+	Provider                   string        `json:"provider"`
+	AuthorizationCheckInterval time.Duration `json:"authorizationCheckInterval"`
+	jwt.RegisteredClaims
+}
+
+func SetZrokCookie(w http.ResponseWriter, domain, email, accessToken, provider string, checkInterval time.Duration, key []byte) {
+	tkn := jwt.NewWithClaims(jwt.SigningMethodHS256, ZrokClaims{
+		Email:                      email,
+		AccessToken:                accessToken,
+		Provider:                   provider,
+		AuthorizationCheckInterval: checkInterval,
+	})
+	sTkn, err := tkn.SignedString(key)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("after signing cookie token: %v", err.Error()), http.StatusInternalServerError)
+		return
+	}
+
+	http.SetCookie(w, &http.Cookie{
+		Name:    "zrok-access",
+		Value:   sTkn,
+		MaxAge:  int(checkInterval.Seconds()),
+		Domain:  domain,
+		Path:    "/",
+		Expires: time.Now().Add(checkInterval),
+		//Secure:  true, //When tls gets added have this be configured on if tls
+	})
+}
+
+func basicAuthRequired(w http.ResponseWriter, realm string) {
 	w.Header().Set("WWW-Authenticate", `Basic realm="`+realm+`"`)
 	w.WriteHeader(401)
-	w.Write([]byte("No Authorization\n"))
+	_, _ = w.Write([]byte("No Authorization\n"))
+}
+
+func oauthLoginRequired(w http.ResponseWriter, r *http.Request, shrToken string, pcfg *Config, provider, target string, authCheckInterval time.Duration) {
+	scheme := "https"
+	if pcfg.Oauth != nil && pcfg.Oauth.RedirectHttpOnly {
+		scheme = "http"
+	}
+	http.Redirect(w, r, fmt.Sprintf("%s://%s.%s:%d/%s/login?targethost=%s&checkInterval=%s", scheme, shrToken, pcfg.Oauth.RedirectHost, pcfg.Oauth.RedirectPort, provider, url.QueryEscape(target), authCheckInterval.String()), http.StatusFound)
 }
 
 func resolveService(hostMatch string, host string) string {
-	logrus.Debugf("host = '%v'", host)
 	if hostMatch == "" || strings.Contains(host, hostMatch) {
 		tokens := strings.Split(host, ".")
 		if len(tokens) > 0 {