+++ ui/src/api/types.js
@@ -253,6 +253,9 @@
  * @property {string} backendProxyEndpoint 
  * @property {string} authScheme 
  * @property {module:types.authUser[]} authUsers 
+ * @property {string} oauthProvider 
+ * @property {string[]} oauthEmailDomains 
+ * @property {string} oauthAuthorizationCheckInterval 
  * @property {boolean} reserved 
  */